refactor: StopWatch 계산 로직을 useStopWatch로 분리
