+++ client/src/components/common/StopWatch.tsx
@@ -1,33 +1,4 @@
-import { useRef, useState, useEffect } from 'react';
-
-interface StopWatchProps {
-  isRunning: boolean;
-}
-
-const StopWatch = ({ isRunning }: StopWatchProps) => {
-  const startTimeRef = useRef(0); // ms
-  const [elapsedSeconds, setElapsedSeconds] = useState(0); // s
-  const [accumulatedTime, setAccumulatedTime] = useState(0); // ms
-
-  useEffect(() => {
-    let animationFrameId: number;
-
-    const step = () => {
-      const diffTime = Date.now() - startTimeRef.current;
-      setElapsedSeconds(Math.floor(diffTime / 1000));
-      animationFrameId = requestAnimationFrame(step);
-    };
-
-    if (isRunning) {
-      startTimeRef.current = Date.now() - accumulatedTime;
-      animationFrameId = requestAnimationFrame(step);
-    } else {
-      setAccumulatedTime(startTimeRef.current ? Date.now() - startTimeRef.current : 0);
-    }
-
-    return () => cancelAnimationFrame(animationFrameId);
-  }, [isRunning]);
-
+const StopWatch = ({ elapsedSeconds = 0 }: { elapsedSeconds: number }) => {
   const formatTime = () => {
     const hours = Math.floor(elapsedSeconds / 3600);
     const minutes = Math.floor((elapsedSeconds % 3600) / 60);
